requirements.txt
mxnet-cu101mkl
torch
torchvision
scipy
matplotlib
pandas
sklearn
Cython
progressbar
TensorBoardX
tqdm
pillow
albumentations
apex
torchsummary